added a component (TodoInfo) to remove redundant states
