+++ src/components/TodoList/TodoList.tsx
@@ -1,6 +1,6 @@
-import React, { memo, useState } from 'react';
-import classNames from 'classnames';
+import React, { memo } from 'react';
 import { Todo } from '../../types/Todo';
+import { TodoInfo } from '../TodoInfo/TodoInfo';
 
 type Props = {
   todos: Todo[];
@@ -8,58 +8,13 @@ type Props = {
 };
 
 export const TodoList: React.FC<Props> = memo(({ todos, onDeleteTodo }) => {
-  const [isLoading, setIsLoading] = useState(false);
-  const [selectedTodoId, setSelectedTodoId] = useState(0);
-  const handleDeleteTodo = async (todoId: number) => {
-    setSelectedTodoId(todoId);
-    setIsLoading(true);
-
-    await onDeleteTodo(todoId);
-
-    setIsLoading(false);
-  };
-
   return (
     <section className="todoapp__main" data-cy="TodoList">
       {todos.map((todo) => (
-        <div
-          data-cy="Todo"
-          key={todo.id}
-          className={classNames('todo', {
-            completed: todo.completed,
-          })}
-        >
-          <label className="todo__status-label">
-            <input
-              data-cy="TodoStatus"
-              type="checkbox"
-              className="todo__status"
-              defaultChecked
-            />
-          </label>
-
-          <span data-cy="TodoTitle" className="todo__title">
-            {todo.title}
-          </span>
-          <button
-            type="button"
-            className="todo__remove"
-            data-cy="TodoDeleteButton"
-            onClick={() => handleDeleteTodo(todo.id)}
-          >
-            ×
-          </button>
-
-          <div
-            data-cy="TodoLoader"
-            className={classNames('modal overlay', {
-              'is-active': isLoading && todo.id === selectedTodoId,
-            })}
-          >
-            <div className="modal-background has-background-white-ter" />
-            <div className="loader" />
-          </div>
-        </div>
+        <TodoInfo
+          todo={todo}
+          onDeleteTodo={onDeleteTodo}
+        />
       ))}
     </section>
   );
